add a suffix to cache file names.
Suppose a URL path "/foo/bar" is cached and stored as "foo/bar".
This prevents another URL path "/foo/bar/zot" being cached as
it needs to create "foo/bar" as a directory, but a regular file
has already been there.

To resolve this issue, Storage adds a consistent prefix (".cache")
to the cache file names.